Fix filtering of files without tags
Introduce directory object type to be able to distinguish between files and directories when executing a filtering query. At the moment, only files should be considered because only files can be tagged, so we don't count a row without a tag if an object is a directory.
This commit also introduces a third parameter to the tags() function to ease debugging.
beholdfs_access() does not return ENOENT when the file is hidden. Probably this should be replaced with a check similar to one in beholdfs_getattr(), so hidden files would give a ENOENT.